+++ src/core/alerts/go-premium-alert.ts
@@ -1,14 +1,17 @@
 import { Alert } from 'react-native'
+import { noop } from '../noop/noop'
 
-export const goPremiumAlert = () => {
+export const goPremiumAlert = (
+  onGoPremium: () => void,
+) => {
   Alert.alert('Want longer results?', 'Buy PREMIUM Plan for only $9.99 / month', [
     {
       text: 'Cancel',
-      onPress: () => null,
+      onPress: noop,
     },
     {
       text: 'Go Premium',
-      onPress: () => null,
+      onPress: onGoPremium,
     }
   ])
 }