+++ src/core/premium-service/premium-service-models.ts
@@ -1,0 +1,6 @@
+import { AnyAction, Dispatch } from '@reduxjs/toolkit'
+
+export interface PremiumService {
+  activateOnce: (dispatch?: Dispatch<AnyAction>) => Promise<void>;
+  deactivateOnce: (dispatch?: Dispatch<AnyAction>) => Promise<void>;
+}