+++ src/core/storage/storage.constants.ts
@@ -2,4 +2,5 @@ export enum STORAGE_KEY {
   SEARCH_RESULT = 'searchResult',
   HAPTIC_FEEDBACK_ENABLED = 'hapticFeedbackEnabled',
   NATIVE_SEARCH_ENGINE_ENABLED = 'nativeSearchEngineEnabled',
+  PREMIUM = 'premium',
 }