+++ src/dashboard/helpers/find-possible-words.helper.ts
@@ -10,6 +10,7 @@ import slowa8 from '../../assets/slowa8'
 import slowa9 from '../../assets/slowa9'
 import { DB } from '../../native-db/native-db'
 import { NATIVE_DB_TAG } from '../../native-db/native-db.constants'
+import { longWordsByLength } from './find-possible-long-words.helper'
 
 export const allWordsByLength = ['', '', slowa2, slowa3, slowa4, slowa5, slowa6, slowa7, slowa8, slowa9]
 
@@ -28,7 +29,7 @@ export const findPossibleWords = async (
   const LOGS_WORD_FILTER = ''
 
   const allWords = R.times(
-    (index: number) => allWordsByLength[minLength + index].split(','),
+    (index: number) => (allWordsByLength[minLength + index] ?? longWordsByLength[minLength + index]).split(','),
     maxLength - minLength + 1
   ).flat().reverse()