+++ src/more/more.styled.ts
@@ -2,6 +2,7 @@ import styled from 'styled-components/native'
 import * as R from 'ramda'
 import { TEXT_SIZE } from '../core/text/text.constants'
 import { COLOR } from '../core/colors/colors.constants'
+import { FocusAwareStatusBar } from '../core/focus-aware-status-bar/focus-aware-status-bar'
 
 interface MoreContainerProps {
   topInset: number;
@@ -10,6 +11,7 @@ interface MoreContainerProps {
 export const MoreContainer = styled.View<MoreContainerProps>`
   padding: 10px;
   flex: 1;
+  background-color: ${COLOR.WHITE_SMOKE};
   margin-top: ${R.propOr(0, 'topInset')}px;
 `
 
@@ -21,3 +23,9 @@ export const MoreDisplayText = styled.Text`
 export const MoreActivityIndicatorWrapper = styled.View`
   margin-top: 30px;
 `
+
+export const MoreStatusBar = styled(FocusAwareStatusBar).attrs({
+  backgroundColor: COLOR.WHITE_SMOKE,
+  animated: true,
+  barStyle: 'dark-content',
+})``