+++ src/settings/store/settings.selectors.ts
@@ -5,3 +5,6 @@ export const hapticFeedbackEnabledSelector = (state: RootState) =>
 
 export const nativeSearchEngineEnabledSelector = (state: RootState) =>
   state?.settings?.nativeSearchEngineEnabled ?? true
+
+export const premiumSelector = (state: RootState) =>
+  state?.settings?.premium ?? 0