+++ src/settings/store/settings.slice.ts
@@ -6,11 +6,13 @@ import { STORAGE_KEY } from '../../core/storage/storage.constants'
 export interface SettingsState {
   hapticFeedbackEnabled: boolean;
   nativeSearchEngineEnabled: boolean;
+  premium: number;
 }
 
 const initialState: SettingsState = {
   hapticFeedbackEnabled: true,
   nativeSearchEngineEnabled: true,
+  premium: 0,
 }
 
 export const settingsSlice = createSlice({
@@ -24,12 +26,16 @@ export const settingsSlice = createSlice({
     setNativeSearchEngineEnabledAction: (state, action: PayloadAction<boolean>) => {
       Storage.set(STORAGE_KEY.NATIVE_SEARCH_ENGINE_ENABLED, String(action.payload))
       state.nativeSearchEngineEnabled = action.payload
+    },
+    setPremiumAction: (state, action: PayloadAction<number>) => {
+      Storage.set(STORAGE_KEY.NATIVE_SEARCH_ENGINE_ENABLED, String(action.payload))
+      state.premium = action.payload
     }
   },
 })
 
 export const {
-  setHapticFeedbackEnabledAction, setNativeSearchEngineEnabledAction,
+  setHapticFeedbackEnabledAction, setNativeSearchEngineEnabledAction, setPremiumAction,
 } = settingsSlice.actions
 
 export const settingsReducer = settingsSlice.reducer